+++ Paddle.h
@@ -5,7 +5,7 @@
 
 struct Paddle {
   Paddle(int16_t x, int16_t y);
-  void update(int16_t target_y, int8_t ball_dx, bool should_miss);
+  void update(uint8_t target_y, int8_t ball_dx, bool should_miss);
   void draw(Adafruit_GFX* display) const;
 
   int16_t x, y;
